Support half-star ratings (UI + DB)
Introduce a reusable StarRating component with half-star support and interactive behavior (tap to toggle full/half/clear). Replace ad-hoc star icons across review forms, feeds and restaurant detail with the new component, update quick-rating UI to require confirmation, and add small UX copy tweaks. Add defensive fixes to MapView to avoid unwanted recenter/fitBounds when the user has interacted and to only run fitBounds once per search. Fix RewardModal so the overlay accepts pointer events. Add a DB migration and schema change to convert integer rating columns to numeric(2,1), handling dependent views and adding a constraint to enforce 0.5 steps.

+++ supabase/schema.sql
@@ -243,7 +243,7 @@ CREATE TABLE IF NOT EXISTS public.content_submissions (
   media_kind    text CHECK (media_kind IN ('photo', 'video')),
   has_comment   boolean NOT NULL DEFAULT false,
   comment_length int NOT NULL DEFAULT 0,
-  rating        int CHECK (rating BETWEEN 1 AND 5),
+  rating        numeric(2,1) CHECK (rating BETWEEN 1 AND 5),
 
   place_id      text,
   place_name    text,
